refactor(control): isolate runtime metrics query
